feat: gRPC merged
