+++ complete_control/config/MasterParams.py
@@ -24,7 +24,7 @@ class MasterParams(BaseModel):
     }
     run_paths: RunPaths
 
-    PLOT_AFTER_SIMULATE: bool = False
+    PLOT_AFTER_SIMULATE: bool = True
     USE_CEREBELLUM: bool = False
     GUI_PYBULLET: bool = False
     USE_MUSIC: bool = False
