+++ complete_control/nrp_bullet_engine.py
@@ -6,12 +6,14 @@ import config.paths as project_paths
 import pybullet as p
 import structlog
 from config.plant_config import PlantConfig
-from nrp_core.engines.python_json import EngineScript
+from nrp_core.engines.python_grpc import GrpcEngineScript
 from plant.plant_simulator import PlantSimulator
 from utils_common.profile import Profile
 
+from nrp_protobuf import nrpgenericproto_pb2, wrappers_pb2
 
-class Script(EngineScript):
+
+class Script(GrpcEngineScript):
 
     def __init__(self):
         super().__init__()
@@ -20,8 +22,9 @@ class Script(EngineScript):
         )
 
     def initialize(self):
-        print("PyBullet Engine Server is initializing.")
+        self.log.info("PyBullet Engine Server is initializing.")
         run_timestamp_str = os.getenv("EXEC_TIMESTAMP")
+        self.log.warning(f"run_timestamp_str =<{run_timestamp_str}>")
         self.run_paths = project_paths.RunPaths.from_run_id(run_timestamp_str)
         self.config = PlantConfig.from_runpaths(self.run_paths)
 
@@ -36,24 +39,25 @@ class Script(EngineScript):
         self.rest_profile = Profile()
         self.log.info("PlantSimulator initialized.")
 
-        self._registerDataPack("positions")
-        self._registerDataPack("control_cmd")
+        # joint_pos_rad (datapack<Double>)
+        self._registerDataPack("joint_pos_rad", wrappers_pb2.DoubleValue)
+        proto_wrapper = wrappers_pb2.DoubleValue()
+        proto_wrapper.value = self.config.master_config.experiment.init_joint_angle
+        self._setDataPack("joint_pos_rad", proto_wrapper)
+        # control_cmd (datapack<Double[]>)
+        self._registerDataPack("control_cmd", nrpgenericproto_pb2.ArrayDouble)
+        proto_wrapper = nrpgenericproto_pb2.ArrayDouble()
+        proto_wrapper.array.extend([0.0, 0.0])
+        self._setDataPack("control_cmd", proto_wrapper)
 
-        self.log.info("DataPacks registered.")
-
-        self._setDataPack(
-            "positions",
-            {"joint_pos_rad": self.config.master_config.experiment.init_joint_angle},
-        )
         self.log.info("NRP Bullet Engine: Initialization complete.")
 
     def runLoop(self, timestep):
         self.rest_profile.end()
         if self.step % 50 == 0:
             self.log.debug("[bullet] starting update...")
-        ctrl = self._getDataPack("control_cmd")
-        rate_pos = ctrl["rate_pos"]
-        rate_neg = ctrl["rate_neg"]
+        ctrl = self._getDataPack("control_cmd").array
+        rate_pos, rate_neg = ctrl[0], ctrl[1]
 
         with self.pybullet_profile.time():
             joint_pos_rad, joint_vel, ee_pos, ee_vel = (
@@ -74,7 +78,10 @@ class Script(EngineScript):
                 time_rest=str(self.rest_profile.total_time),
             )
 
-        self._setDataPack("positions", {"joint_pos_rad": joint_pos_rad})
+        datapack = wrappers_pb2.DoubleValue()
+        datapack.value = joint_pos_rad
+        self._setDataPack("joint_pos_rad", datapack)
+
         self.rest_profile.start()
 
     def shutdown(self):
