+++ complete_control/plant/plant_utils.py
@@ -105,7 +105,8 @@ def compute_spike_rate(
     # Using a more direct iteration for potentially better performance with many spikes
     # than converting to numpy array first for just filtering.
     count = 0
-    for t, _ in spikes:
+    # TODO fix for music...
+    for t in spikes:
         if time_start <= t < time_end:
             count += 1
