+++ nrp_simulation_config_nest_docker_compose.json
@@ -8,7 +8,7 @@
             "EngineType": "python_grpc",
             "EngineName": "bullet_simulator",
             "ServerAddress":"0.0.0.0:1234",
-            "PythonFileName": "complete_control/nrp_bullet_engine_bk.py",
+            "PythonFileName": "complete_control/nrp_bullet_engine.py",
 			"ProtobufPackages": ["Wrappers", "NrpGenericProto"],
             "EngineTimestep" : 0.0001
         },
@@ -16,7 +16,7 @@
             "EngineType": "python_grpc",
 			"EngineName": "nest_client",
 			"ServerAddress":"0.0.0.0:1235",
-			"PythonFileName": "complete_control/nrp_neural_engine_bk.py",
+			"PythonFileName": "complete_control/nrp_neural_engine.py",
 			"ProtobufPackages": ["Wrappers", "NrpGenericProto"],
             "EngineTimestep" : 0.0001
 		}
